+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Float values not allowed for addMatomoMethodCall (#30)
+
 ## [2.1.0] - 2024-04-19
 
 ### Added